Ratchet mypy on modules used by Lyft's GitHub ingestion
Lyft has additional changes planned to its GitHub code in the near term,
so ratchet up the strictness flags to catch any errors.
Note that we leave `disallow_any_generics = false`;
setting it to `true` would be more strict
but it's not clear it's yet worth it to add those types,
vs switching to e.g. TypedDict.

The `GraphJob.run_from_json` and `GraphJob.run_from_json_file`
methods lose the default value of `None` for their `parameters` argument;
this appears to be a breaking change for API callers not passing that argument.
However, the parameters argument is passed through to a dict.update call,
which does not accept `None` as an input,
so any callers not passing that argument would have failed at runtime anyways.