Loosen typing on `GraphJob` json `file_path`s
The `file_path` parameter is passed directly to `open`,
which supports `typing.Union[str, bytes, os.PathLike]`,
and internally at Lyft we always pass a string for this parameter.

We need to parse out a shortname as a string from the path,
so we can't expand this type to include bytes;
only expand to be a Union of str and Path.